rename : 회원 package 구조 변경
