+++ src/main/java/com/crude/travelcrew/domain/crew/model/dto/CrewListRes.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 import com.crude.travelcrew.domain.crew.model.constants.CrewStatus;
 import com.crude.travelcrew.domain.crew.model.entity.Crew;
-import com.crude.travelcrew.domain.member.entity.Member;
+import com.crude.travelcrew.domain.member.model.entity.Member;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
