+++ src/main/java/com/crude/travelcrew/domain/crew/model/entity/Crew.java
@@ -20,7 +20,7 @@ import com.crude.travelcrew.domain.crew.model.constants.CrewGender;
 import com.crude.travelcrew.domain.crew.model.constants.CrewStatus;
 import com.crude.travelcrew.domain.crew.model.dto.CrewReq;
 import com.crude.travelcrew.domain.crew.model.dto.CrewRes;
-import com.crude.travelcrew.domain.member.entity.Member;
+import com.crude.travelcrew.domain.member.model.entity.Member;
 import com.crude.travelcrew.global.entity.BaseTime;
 
 import lombok.Getter;
