+++ src/main/java/com/crude/travelcrew/domain/crew/repository/CrewRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import com.crude.travelcrew.domain.crew.model.entity.Crew;
-import com.crude.travelcrew.domain.member.entity.Member;
+import com.crude.travelcrew.domain.member.model.entity.Member;
 
 public interface CrewRepository extends JpaRepository<Crew, Long> {
 	@Query(
