+++ src/main/java/com/crude/travelcrew/domain/crew/repository/CrewScrapRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.crude.travelcrew.domain.crew.model.entity.Crew;
 import com.crude.travelcrew.domain.crew.model.entity.CrewScrap;
-import com.crude.travelcrew.domain.member.entity.Member;
+import com.crude.travelcrew.domain.member.model.entity.Member;
 
 public interface CrewScrapRepository extends JpaRepository<CrewScrap, Long> {
 	Optional<CrewScrap> findByMemberAndCrew(Member member, Crew crew);
