--- src/main/java/com/crude/travelcrew/domain/member/constants/MemberRole.java
@@ -1,8 +1,0 @@
-package com.crude.travelcrew.domain.member.constants;
-
-import lombok.Getter;
-
-@Getter
-public enum MemberRole {
-	ADMIN, MANAGER, USER;
-}
