+++ src/main/java/com/crude/travelcrew/domain/member/model/constants/MemberRole.java
@@ -1,0 +1,8 @@
+package com.crude.travelcrew.domain.member.model.constants;
+
+import lombok.Getter;
+
+@Getter
+public enum MemberRole {
+	ADMIN, MANAGER, USER
+}
