+++ src/main/java/com/crude/travelcrew/domain/member/repository/MemberProfileRepository.java
@@ -2,7 +2,7 @@ package com.crude.travelcrew.domain.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.crude.travelcrew.domain.member.entity.MemberProfile;
+import com.crude.travelcrew.domain.member.model.entity.MemberProfile;
 
 public interface MemberProfileRepository extends JpaRepository<MemberProfile, Long> {
 }
