+++ src/main/java/com/crude/travelcrew/domain/member/service/MemberService.java
@@ -7,13 +7,13 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.crude.travelcrew.domain.member.constants.MemberRole;
-import com.crude.travelcrew.domain.member.constants.MemberStatus;
-import com.crude.travelcrew.domain.member.constants.ProviderType;
-import com.crude.travelcrew.domain.member.dto.SignUpReq;
-import com.crude.travelcrew.domain.member.dto.SignUpRes;
-import com.crude.travelcrew.domain.member.entity.Member;
-import com.crude.travelcrew.domain.member.entity.MemberProfile;
+import com.crude.travelcrew.domain.member.model.constants.MemberRole;
+import com.crude.travelcrew.domain.member.model.constants.MemberStatus;
+import com.crude.travelcrew.domain.member.model.constants.ProviderType;
+import com.crude.travelcrew.domain.member.model.dto.SignUpReq;
+import com.crude.travelcrew.domain.member.model.dto.SignUpRes;
+import com.crude.travelcrew.domain.member.model.entity.Member;
+import com.crude.travelcrew.domain.member.model.entity.MemberProfile;
 import com.crude.travelcrew.domain.member.repository.MemberProfileRepository;
 import com.crude.travelcrew.domain.member.repository.MemberRepository;
 import com.crude.travelcrew.global.security.JwtProvider;
