+++ src/main/java/com/crude/travelcrew/domain/report/model/entity/Report.java
@@ -10,7 +10,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import com.crude.travelcrew.domain.member.entity.Member;
+import com.crude.travelcrew.domain.member.model.entity.Member;
 import com.crude.travelcrew.global.entity.BaseTime;
 
 import lombok.AccessLevel;
